// packages/server/src/@types/db.schema.ts
export type job_experience_level = 'entry' | 'executive' | 'internship' | 'lead' | 'manager' | 'mid' | 'senior';
export type job_pay_type = 'bi_weekly' | 'daily' | 'hourly' | 'monthly' | 'weekly' | 'yearly';
export type job_remote_status = 'hybrid' | 'onsite' | 'remote';
export type job_type = 'contract' | 'full_time' | 'internship' | 'part_time' | 'remote' | 'temporary' | 'volunteer';

// Table jobs
export interface Jobs {
  id: string;
  title: string;
  description: string | null;
  url: string;
  date_posted: Date | null;
  remote_status: job_remote_status;
  location: string | null;
  company: string | null;
  experience_level: job_experience_level;
  type: job_type;
  pay_min: number | null;
  pay_max: number | null;
  pay_type: job_pay_type | null;
  archived: boolean;
  created_by: string;
  created_at: Date;
  updated_at: Date;
}
export interface JobsInput {
  id?: string;
  title: string;
  description?: string | null;
  url: string;
  date_posted?: Date | null;
  remote_status: job_remote_status;
  location?: string | null;
  company?: string | null;
  experience_level: job_experience_level;
  type: job_type;
  pay_min?: number | null;
  pay_max?: number | null;
  pay_type?: job_pay_type | null;
  archived?: boolean;
  created_by: string;
  created_at?: Date;
  updated_at?: Date;
}
// ...
export interface Users {
  /** Unique identifier */
  id: string;
  /** Email address */
  email: string;
  /** Password */
  password: string;
  /** Name */
  name: string;
  /** Creation date */
  created_at: Date;
  /** Last update date */
  updated_at: Date;
}